==> courses/No05-2025-OReilly-LLM/Week 1/01-similarity-embeddings.ipynb <==
# Load data

import glob
import os

data = []
for n in glob.glob("un/TXT/Session 77 - 2022/*.txt"):
    data.append({"country": os.path.basename(n.replace("_77_2022.txt", "")), "text": open(n).read() })

import pandas as pd

pd.set_option('display.max_colwidth', 500)
df = pd.DataFrame(data)
df

# Sentence segmentation

import spacy

!python -m spacy download en_core_web_sm

nlp = spacy.load("en_core_web_sm")

from tqdm.auto import tqdm

# runs < 1min
sentences = []
for text in tqdm(df["text"]):
    doc = nlp(text)
    for sentence in doc.sents:
        sentences.append(str(sentence).strip())

sentences[0:20]

len(sentences)

open("sentences.txt", "w").write("@@@".join(sentences))

sentences = open("sentences.txt").read().split("@@@")

# Encode sentences

!pip install sentence_transformers -U

from sentence_transformers import SentenceTransformer
model = SentenceTransformer('multi-qa-MiniLM-L6-cos-v1')

!pip install huggingface_hub -U

# can take a minute or two depending on CPU/GPU configuration
sembeddings = model.encode(sentences, show_progress_bar=True, normalize_embeddings=True)

len(sembeddings)

sembeddings.shape

import numpy as np
with open("sentences-mqa.npy", "wb") as f:
    np.save(f, sembeddings)

Many more models are available on Hugging Face.

Benchmark of models: https://huggingface.co/spaces/mteb/leaderboard

Search for all sentence similarity models: https://huggingface.co/models?pipeline_tag=sentence-similarity&sort=trending

# new superfast alternative using ModelVec, new in 3.4, speedup 400x CPU to 25x GPU:
# model = SentenceTransformer("minishlab/potion-base-8M", device="cpu")

## Alternative Model

# option: truncate_dim=dimensions
# option for cpu: backend="openvino"
model2 = SentenceTransformer('mixedbread-ai/mxbai-embed-large-v1')

# can take a minute or two depending on CPU/GPU configuration
sembeddings2 = model2.encode(sentences, show_progress_bar=True, 
                             normalize_embeddings=True)

sembeddings2.shape

# if we want, we could now quantize the embeddings to save space and add performance:
from sentence_transformers.quantization import quantize_embeddings
binary_embeddings2 = quantize_embeddings(sembeddings2, precision="ubinary")
binary_embeddings2.shape

with open("sentences-mbread.npy", "wb") as f:
    np.save(f, sembeddings2)

## One more alternative

model3 = SentenceTransformer("NovaSearch/stella_en_1.5B_v5", trust_remote_code=True)

sembeddings3 = model3.encode(sentences, show_progress_bar=True, normalize_embeddings=True)

sembeddings3.shape

with open("sentences-stella.npy", "wb") as f:
    np.save(f, sembeddings3)

# Retrieval

def search(query, text, corpus_embeddings, model, query_prompt_name=None, top=20):
    # code query to restrict search space
    question_embedding = model.encode(query, normalize_embeddings=True, prompt_name=query_prompt_name)
    
    # Determine similarity (vectors are normalized)
    sim = model.similarity(question_embedding, corpus_embeddings)[0].numpy() 
    # Alternative: sim = np.dot(corpus_embeddings, question_embedding)
    
    # Get most similar top_k by sorting
    hits = [ { "text": text[i], "score": sim[i] } 
                     for i in sim.argsort()[::-1][0:top] ]
    
    # Return as dataframe
    return pd.DataFrame(hits)

pd.set_option('display.max_colwidth', 0)

search("The climate crisis is worse for poorer countries.", sentences, sembeddings, model)

search("The climate crisis is worse for poorer countries.", sentences, sembeddings2, model2)

search("The climate crisis is worse for poorer countries.", sentences, sembeddings2, model2,
       query_prompt_name="query")

search("The climate crisis is worse for poorer countries.", sentences, sembeddings3, model3)

search("The climate crisis is worse for poorer countries.", sentences, sembeddings3, model3, 
       query_prompt_name="s2p_query")
